vulnerable_chatbot/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pypdf
python-docx